network_probe/requirements.txt
PyQt6>=6.5.0
dnspython>=2.4.0
requests>=2.31.0
pyinstaller>=6.0.0
